+++ index.ts
@@ -50,6 +50,7 @@ export { default as NewsMessageVideo } from './src/structures/NewsMessageVideo';
 export { default as RadioStation } from './src/structures/RadioStation';
 export { default as Stats } from './src/structures/Stats';
 export { default as Tournament } from './src/structures/Tournament';
+export { default as TournamentManager } from './src/structures/TournamentManager';
 export { default as TournamentWindow } from './src/structures/TournamentWindow';
 export { default as BaseFriendMessage } from './src/structures/friend/BaseFriendMessage';
 export { default as BasePendingFriend } from './src/structures/friend/BasePendingFriend';
