+++ src/structures/user/User.ts
@@ -130,7 +130,7 @@ class User extends Base {
    * @throws {EpicgamesAPIError}
    */
   public async getEventTokens(): Promise<EventTokens | undefined> {
-    return (await this.client.getEventTokens(this.id))[0];
+    return (await this.client.tournaments.getEventTokens(this.id))[0];
   }
 
   /**
